+++ recurrent_OKFTJ/src/servlet/WelcomeServlet.java
@@ -8,6 +8,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+
+import bo.WelcomeLogic;
+import model.Terminal;
 
 @WebServlet("/WelcomeServlet")
 public class WelcomeServlet extends HttpServlet {
@@ -15,9 +19,18 @@ public class WelcomeServlet extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
-		//QRコードログイン
-		//		String qrCode = request.getParameter("store");
-		//		System.out.println(qrCode);
+		request.setCharacterEncoding("utf-8");
+		//店舗＿席情報登録、注文ＩＤ取得(滝本)
+		{
+			String store_seatId = request.getParameter("store_seatId");
+
+			Terminal terminal = new Terminal(store_seatId);
+			WelcomeLogic welcomeL = new WelcomeLogic();
+			welcomeL.execute(terminal);
+
+			HttpSession session = request.getSession();
+			session.setAttribute("terminal", terminal);
+		}
 
 		RequestDispatcher disp = request.getRequestDispatcher("/WEB-INF/jsp/welcome.jsp");
 		disp.forward(request, response);
@@ -29,8 +42,6 @@ public class WelcomeServlet extends HttpServlet {
 
 		RequestDispatcher disp = request.getRequestDispatcher("/WEB-INF/jsp/top.jsp");
 		disp.forward(request, response);
-
-		doGet(request, response);
 	}
 
 }
